StepTalk Smalltalk compiler: int->NSInteger transition
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/steptalk/trunk@36658 72102866-910b-0410-8b05-ffd578937521

+++ Languages/Smalltalk/NSNumber+additions.h
@@ -35,6 +35,6 @@
 - (BOOL)isTrue;
 - (BOOL)isFalse;
 
-- to:(int)number do:(STBlock *)block;
-- to:(int)number step:(int)step do:(STBlock *)block;
+- to:(NSInteger)number do:(STBlock *)block;
+- to:(NSInteger)number step:(NSInteger)step do:(STBlock *)block;
 @end
